+++ controllers/apply-next/simple/processor.js
@@ -4,7 +4,7 @@ const debug = require('debug')('tnlcf:a4a');
 const features = require('config').get('features');
 
 const { fromDateParts } = require('../../../modules/dates');
-const salesforceService = require('../../../services/salesforce');
+const salesforceService = require('../lib/salesforce');
 
 function salesforceApplication(application) {
     function dateFormat(dt) {
